change wsm_client_create() API, take wl_client as parameter
wsm_client_create() is supposed to be called with a client
file descriptor as a parameter. However, a compositor
implementing libwsm would not be able to obtain such a fd,
as this is an internal private information -not exposed by
libwayland.

On the other hand, libwayland already provides a
wl_client_get_credentials() function, which provides 100%
of the functionality we need.

As a result :
- replace the "int" parameter with a "wl_client" one ;
- wayland-server library becomes a hard dependency ;
- adapt the base testcase -it still passes.